requirements.txt
google-genai
pandas
fastapi
uvicorn
dotenv
pydantic
pymysql
SQLAlchemy
pyJWT
